backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
openai==1.3.0
pydantic==1.10.8
python-dotenv==1.0.0
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
Pillow==10.0.0
langchain==0.0.267
langchain-openai==0.0.2 